+++ Sources/RichTextKit/Images/Image+RichText.swift
@@ -25,6 +25,17 @@ public extension Image {
     static let richTextAlignmentJustified = symbol("text.justify")
     static let richTextAlignmentLeft = symbol("text.alignleft")
     static let richTextAlignmentRight = symbol("text.alignright")
+    
+    static func richTextAlignment(_ alignment: NSTextAlignment) -> Image {
+        switch alignment {
+        case .center: richTextAlignmentCenter
+        case .justified: richTextAlignmentJustified
+        case .left: richTextAlignmentLeft
+        case .natural: richTextAlignmentLeft
+        case .right: richTextAlignmentRight
+        @unknown default: richTextUnknownValueType
+        }
+    }
 
     static let richTextColorBackground = symbol("highlighter")
     static let richTextColorForeground = symbol("character")
@@ -60,6 +71,8 @@ public extension Image {
 
     static let richTextSuperscriptDecrease = symbol("textformat.subscript")
     static let richTextSuperscriptIncrease = symbol("textformat.superscript")
+    
+    static let richTextUnknownValueType = symbol("questionmark")
 }
 
 public extension Image {